modularizing drive auth api and client-api
