+++ example/data.js
@@ -1,0 +1,1 @@
+[{"fileID":"0AF6S8E0_wBt-Uk9PVA","fileName":"Root","children":[{"fileID":"0B16S8E0_wBt-Mjd1a0pOWmlUOFk","fileName":"testing","fileIcon":"https://ssl.gstatic.com/docs/doclist/images/icon_11_collection_list.png","fileLink":"https://docs.google.com/folderview?id=0B16S8E0_wBt-Mjd1a0pOWmlUOFk&usp=drivesdk","fileType":"application/vnd.google-apps.folder","children":[{"fileID":"0B16S8E0_wBt-OXNQcW90SVZXaEE","fileName":"aziz2.html","fileIcon":"https://ssl.gstatic.com/docs/doclist/images/icon_10_generic_list.png","fileLink":"https://docs.google.com/file/d/0B16S8E0_wBt-OXNQcW90SVZXaEE/edit?usp=drivesdk","fileType":"text/html","children":[]},{"fileID":"0B16S8E0_wBt-ZkNWZl9fRnU5TlE","fileName":"testing2","fileIcon":"https://ssl.gstatic.com/docs/doclist/images/icon_11_collection_list.png","fileLink":"https://docs.google.com/folderview?id=0B16S8E0_wBt-ZkNWZl9fRnU5TlE&usp=drivesdk","fileType":"application/vnd.google-apps.folder","children":[{"fileID":"0B16S8E0_wBt-NEdGMkpfdUktdU0","fileName":"ChatBuilder 2.html","fileIcon":"https://ssl.gstatic.com/docs/doclist/images/icon_10_generic_list.png","fileLink":"https://docs.google.com/file/d/0B16S8E0_wBt-NEdGMkpfdUktdU0/edit?usp=drivesdk","fileType":"text/html","children":[]},{"fileID":"0B16S8E0_wBt-Rl9jeHRnMVZiaHM","fileName":"testing3","fileIcon":"https://ssl.gstatic.com/docs/doclist/images/icon_11_collection_list.png","fileLink":"https://docs.google.com/folderview?id=0B16S8E0_wBt-Rl9jeHRnMVZiaHM&usp=drivesdk","fileType":"application/vnd.google-apps.folder","children":[{"fileID":"0B16S8E0_wBt-Q2dmVldrc2lqZ2M","fileName":"google-oauth-server.js","fileIcon":"https://ssl.gstatic.com/docs/doclist/images/icon_10_generic_list.png","fileLink":"https://docs.google.com/file/d/0B16S8E0_wBt-Q2dmVldrc2lqZ2M/edit?usp=drivesdk","fileType":"application/x-javascript","children":[]}]}]}]}]}]
