+++ server/auth/auth-router.js
@@ -1,7 +1,6 @@
 var express = require('express');
 var dropboxRouter = require('./dropbox-router.js');
 var driveRouter = require('./drive-router.js')
-var secrets = require('./secrets/drive.secret')
 var authRouter = express.Router();
 
 authRouter.use('/dropbox', dropboxRouter);
